+++ services/auth/auth.go
@@ -8,11 +8,11 @@ import (
 
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/jackc/pgx/v5"
+	"github.com/piquel-fr/api/config"
 	"github.com/piquel-fr/api/database"
 	"github.com/piquel-fr/api/database/repository"
-	"github.com/piquel-fr/api/errors"
 	"github.com/piquel-fr/api/services/auth/oauth"
-	"github.com/piquel-fr/api/services/config"
+	"github.com/piquel-fr/api/utils/errors"
 )
 
 func InitAuthService() {
