+++ utils/github/github.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 
 	"github.com/google/go-github/v74/github"
-	"github.com/piquel-fr/api/errors"
-	"github.com/piquel-fr/api/services/config"
+	"github.com/piquel-fr/api/config"
+	"github.com/piquel-fr/api/utils/errors"
 )
 
 var Client *github.Client
